Moved body description of booking dialog into BookingDialog
