+++ src/api.ts
@@ -139,5 +139,9 @@ async function fetchJsonDataAsync<T>(query: string): Promise<{ data: T }> {
   if (!data) {
     throw new Error("Response error");
   }
+  function sleep(ms: number) {
+    return new Promise((resolve) => setTimeout(resolve, ms));
+  }
+  await sleep(500);
   return { data };
 }
