+++ src/components/dialogs/BookingDialog.tsx
@@ -7,16 +7,27 @@ import { useAppDispatch } from "../../redux/hooks";
 import * as Api from "../../api";
 import * as ConnectionErrorSlice from "../../redux/connectionErrorSlice";
 
-import BookingDialogBody from "./BookingDialogBody";
+import LoadingDataWrapper from "./LoadingDataWrapper";
+import DescriptionRow from "./DescriptionRow";
+import RoomContainer from "./RoomContainer";
+import GuestRow from "./GuestRow";
 
 type Props = {
   tileId?: string,
   bookingId?: string
 };
 
+const initialData: Api.BookingData = {
+  id: "",
+  name: "",
+  from: "",
+  to: "",
+  rooms: []
+};
+
 function BookingDialog(props: Props): JSX.Element {
   const dispatch = useAppDispatch();
-  const [bookingData, setBookingData] = useState<Api.BookingData>();
+  const [bookingData, setBookingData] = useState<Api.BookingData>(initialData);
 
   const bookingTitle = getTitleFromBookingData(bookingData);
 
@@ -25,20 +36,34 @@ function BookingDialog(props: Props): JSX.Element {
   return (
     <div className="scrollable">
       <DialogHeader title={`Prenotazione ${bookingTitle}`} />
-      <BookingDialogBody data={bookingData} />
+      <LoadingDataWrapper isLoaded={bookingData.id.length > 0}>
+        <DescriptionRow name="Dal" value={new Date(bookingData.from).toLocaleDateString()} />
+        <DescriptionRow name="Al" value={new Date(bookingData.to).toLocaleDateString()} />
+        <h3 className="sub-header">Stanze</h3>
+        <hr />
+        <div className="rooms-container">
+          {bookingData.rooms.map(room => (
+            <RoomContainer key={room.id} data={room}>
+              {room.guests.map(guest => (
+                <GuestRow key={guest.id} data={guest}/>
+              ))}
+            </RoomContainer>
+          ))}
+        </div>
+      </LoadingDataWrapper>
     </div>
   );
 }
 
-function getTitleFromBookingData(data: Api.BookingData | undefined): string {
-  return data === undefined ? "" : `#${data.id} (${data.name})`;
+function getTitleFromBookingData(data: Api.BookingData): string {
+  return data.id.length === 0 ? "" : `#${data.id} (${data.name})`;
 }
 
 function useFetchDataEffect(
   tileId: string | undefined,
   bookingId: string | undefined,
   dispatch: Dispatch<AnyAction>,
-  setBookingData: React.Dispatch<React.SetStateAction<Api.BookingData | undefined>>
+  setBookingData: React.Dispatch<React.SetStateAction<Api.BookingData>>
 ): void {
   useEffect(() => {
     async function fetchData() {
